[MOVE] LeagueTeam, LeagueTeamPlayer domain 으로 이동 #157
